Reorganizes the Ranges to be under the rules component
Ranges are more of an artifact of communicating possible rules/actions, than they are related to the state of the game.
