+++ src/main/java/pro/trevor/tankgame/rule/definition/player/IPlayerRule.java
@@ -1,7 +1,7 @@
 package pro.trevor.tankgame.rule.definition.player;
 
+import pro.trevor.tankgame.rule.definition.range.TypeRange;
 import pro.trevor.tankgame.state.State;
-import pro.trevor.tankgame.util.range.TypeRange;
 
 public interface IPlayerRule<T> {
