+++ src/main/java/pro/trevor/tankgame/rule/definition/player/TimedPlayerActionRule.java
@@ -2,12 +2,12 @@ package pro.trevor.tankgame.rule.definition.player;
 
 import org.json.JSONObject;
 import pro.trevor.tankgame.Main;
+import pro.trevor.tankgame.rule.definition.range.TypeRange;
 import pro.trevor.tankgame.rule.type.ICooldownPlayerElement;
 import pro.trevor.tankgame.state.State;
 import pro.trevor.tankgame.util.IJsonObject;
 import pro.trevor.tankgame.util.function.IVarTriConsumer;
 import pro.trevor.tankgame.util.function.IVarTriPredicate;
-import pro.trevor.tankgame.util.range.TypeRange;
 
 import java.util.Arrays;
 import java.util.function.Function;
