+++ src/main/java/pro/trevor/tankgame/rule/definition/range/BooleanRange.java
@@ -1,4 +1,4 @@
-package pro.trevor.tankgame.util.range;
+package pro.trevor.tankgame.rule.definition.range;
 
 import java.util.Set;
 import java.util.stream.Collectors;
