+++ src/main/java/pro/trevor/tankgame/rule/definition/range/DiscreteIntegerRange.java
@@ -1,4 +1,4 @@
-package pro.trevor.tankgame.util.range;
+package pro.trevor.tankgame.rule.definition.range;
 
 import java.util.HashSet;
 import java.util.Set;
