+++ src/main/java/pro/trevor/tankgame/rule/definition/range/IntegerRange.java
@@ -1,4 +1,4 @@
-package pro.trevor.tankgame.util.range;
+package pro.trevor.tankgame.rule.definition.range;
 
 public class IntegerRange extends BaseRange<Integer> {
