+++ src/main/java/pro/trevor/tankgame/rule/definition/range/MovePositionRange.java
@@ -1,9 +1,8 @@
-package pro.trevor.tankgame.rule.impl.shared.range;
+package pro.trevor.tankgame.rule.definition.range;
 
 import pro.trevor.tankgame.state.State;
 import pro.trevor.tankgame.state.board.Position;
 import pro.trevor.tankgame.state.board.unit.GenericTank;
-import pro.trevor.tankgame.state.range.FunctionVariableRange;
 import pro.trevor.tankgame.util.Util;
 
 import java.util.HashSet;
