+++ src/main/java/pro/trevor/tankgame/rule/definition/range/VariableTypeRange.java
@@ -1,7 +1,6 @@
-package pro.trevor.tankgame.state.range;
+package pro.trevor.tankgame.rule.definition.range;
 
 import pro.trevor.tankgame.state.State;
-import pro.trevor.tankgame.util.range.DiscreteTypeRange;
 
 public interface VariableTypeRange<T, U> extends DiscreteTypeRange<U> {
